+++ CHANGELOG.md
@@ -1,3 +1,85 @@
+# 1.17.0 / 2025-06-02
+
+* [MAINTENANCE] Update Datadog SDK to 2.22.0.
+  See [#376](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/376)
+* [MAINTENANCE] Use SDK snapshot version for instrumented test.
+  See [#372](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/372)
+
+# 1.17.0-rc1 / 2025-04-28
+
+* [FEATURE] Add Kotlin Compiler Plugin environment.
+  See [#314](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/314)
+* [FEATURE] Add `ComposeTagTransformer` to inject `datadog` modifier call.
+  See [#321](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/321)
+* [FEATURE] Register plugin to Kotlin compile options.
+  See [#322](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/322)
+* [FEATURE] Implement annotation mode for instrumentation.
+  See [#342](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/342)
+* [FEATURE] Implement `NavHost` instrumentation transformer.
+  See [#329](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/329)
+* [FEATURE] Implement Kotlin Compiler Plugin DSL configuration.
+  See [#330](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/330)
+* [FEATURE] Add `semantics role` for `Image`, `Icon`, and `AsyncImage`.
+  See [#334](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/334)
+* [FEATURE] Use composable function name as the semantics tag.
+  See [#337](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/337)
+* [IMPROVEMENT] Filter out `ksp` and `KaptGenerateStubsTask` from `freeCompilerArgs`.
+  See [#335](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/335)
+* [IMPROVEMENT] Fix gitlab-ci instrumented task.
+  See [#331](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/331)
+* [IMPROVEMENT] Enhance Compiler plugin binary backward compatibility.
+  See [#345](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/345)
+* [IMPROVEMENT] Add precise error message for missing dependencies.
+  See [#361](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/361)
+* [IMPROVEMENT] Only warn missing dependencies instead of throwing errors.
+  See [#368](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/368)
+* [IMPROVEMENT] Fix tag injection when `Modifier` is absent.
+  See [#357](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/357)
+* [IMPROVEMENT] Do not show Android project error if KCP is enabled.
+  See [#359](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/359)
+* [IMPROVEMENT] Use Sonatype snapshot URL and fix compile warnings.
+  See [#367](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/367)
+* [MAINTENANCE] Bump Kotlin version and embeddable to `2.0.21`.
+  See [#339](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/339)
+* [MAINTENANCE] Separate Compose extension to register respectively.
+  See [#332](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/332)
+* [MAINTENANCE] Merge `develop` into `feature/kotlin-compiler-plugin`.
+  See [#355](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/355)
+* [MAINTENANCE] Merge Compose instrumentation configuration into one.
+  See [#358](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/358)
+* [MAINTENANCE] Fix instrumented test task timeout issue.
+  See [#362](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/362)
+* [MAINTENANCE] Dev: Start next development iteration.
+  See [#351](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/351)
+* [MAINTENANCE] Datadog Gradle Plugin: Update to version `1.16.0`.
+  See [#354](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/354)
+* [MAINTENANCE] Merge `release/1.16.0` into `develop`.
+  See [#353](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/353)
+* [MAINTENANCE] Add compilation unit test for `modifier` injection.
+  See [#323](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/323)
+* [MAINTENANCE] Setup compiling test for `NavHost` injection.
+  See [#328](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/328)
+* [MAINTENANCE] Merge `develop` into `feature/kotlin-compiler-plugin`.
+  See [#317](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/317)
+* [MAINTENANCE] Kotlin Compiler Plugin: Add `instrumented` test module.
+  See [#315](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/315)
+* [MAINTENANCE] Merge `feature/kotlin-compiler-plugin` into `develop`.
+  See [#360](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/360)
+* [MAINTENANCE] Start to publish SNAPSHOT from `develop` branch.
+  See [#366](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/366)
+* [MAINTENANCE] CodeQL: Update Ubuntu version from `20.04` to `latest`.
+  See [#364](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/364)
+* [MAINTENANCE] Merge `develop` into `feature/kotlin-compiler-plugin`.
+  See [#365](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/365)
+* [MAINTENANCE] Update Compose Instrumentation functions references.
+  See [#363](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/363)
+* [MAINTENANCE] Merge `develop` into `feature/kotlin-compiler-plugin`.
+  See [#336](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/336)
+* [MAINTENANCE] Add Instrumented test for view tracking instrumentation.
+  See [#341](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/341)
+* [MAINTENANCE] Update the package name of instrumentation API.
+  See [#343](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/343)
+
 # 1.16.0 / 2025-04-07
 
 * [IMPROVEMENT] Log message if there are no files to upload. See [#309](https://github.com/DataDog/dd-sdk-android-gradle-plugin/pull/309)